# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Onyx Meridian Enterprise OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a single agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Not found
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent status or config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, name, role, department, status, autonomyLevel, utilizationPct, slaPerformance, costPerDay, taskCount]
      properties:
        id:
          type: string
        name:
          type: string
        employeeId:
          type: string
        role:
          type: string
        department:
          type: string
        manager:
          type: ["string", "null"]
        status:
          type: string
          enum: [deployed, draft, configured, suspended, retired]
        autonomyLevel:
          type: string
          enum: [full, supervised, assisted]
        utilizationPct:
          type: number
        slaPerformance:
          type: number
        costPerDay:
          type: number
        taskCount:
          type: integer
        skills:
          type: array
          items:
            type: string
        businessUnitId:
          type: string
    AgentDetail:
      type: object
      required: [id, name, role, department, status, autonomyLevel, utilizationPct, slaPerformance, costPerDay, taskCount, skills, tools, permissions, budget, reasoning, memoryUsage, tokenUsage, latency, accuracy, hallucination, policyCompliance]
      properties:
        id:
          type: string
        name:
          type: string
        employeeId:
          type: string
        role:
          type: string
        department:
          type: string
        manager:
          type: ["string", "null"]
        status:
          type: string
          enum: [deployed, draft, configured, suspended, retired]
        autonomyLevel:
          type: string
          enum: [full, supervised, assisted]
        utilizationPct:
          type: number
        slaPerformance:
          type: number
        costPerDay:
          type: number
        taskCount:
          type: integer
        skills:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
        budget:
          type: number
        reasoning:
          type: string
        memoryUsage:
          type: number
        tokenUsage:
          type: integer
        latency:
          type: number
        accuracy:
          type: number
        hallucination:
          type: number
        policyCompliance:
          type: number
        businessUnitId:
          type: string
        recentTasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
    AgentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [deployed, draft, configured, suspended, retired]
        autonomyLevel:
          type: string
          enum: [full, supervised, assisted]
...
    Task:
      type: object
      required: [id, title, status, priority, owner, linkedKpi, dueDate, progress, businessUnitId]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [todo, in-progress, done, blocked]
        priority:
          type: string
          enum: [p1, p2, p3]
        owner:
          type: string
        linkedKpi:
          type: string
        dueDate:
          type: string
        progress:
          type: integer
        businessUnitId:
          type: string
        aiGenerated:
          type: boolean
        department:
          type: ["string", "null"]
...