# lib/api-zod/src/generated/types/agentDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onyx Meridian Enterprise OS API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentDetailAutonomyLevel } from './agentDetailAutonomyLevel';
import type { AgentDetailStatus } from './agentDetailStatus';
import type { Task } from './task';

export interface AgentDetail {
  id: string;
  name: string;
  employeeId?: string;
  role: string;
  department: string;
  /** @nullable */
  manager?: string | null;
  status: AgentDetailStatus;
  autonomyLevel: AgentDetailAutonomyLevel;
  utilizationPct: number;
  slaPerformance: number;
  costPerDay: number;
  taskCount: number;
  skills: string[];
  tools: string[];
  permissions: string[];
  budget: number;
  reasoning: string;
  memoryUsage: number;
  tokenUsage: number;
  latency: number;
  accuracy: number;
  hallucination: number;
  policyCompliance: number;
  businessUnitId?: string;
  recentTasks?: Task[];
}