# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Onyx Meridian Enterprise OS API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentUpdateAutonomyLevel } from './agentUpdateAutonomyLevel';
import type { AgentUpdateStatus } from './agentUpdateStatus';

export interface AgentUpdate {
  status?: AgentUpdateStatus;
  autonomyLevel?: AgentUpdateAutonomyLevel;
}